GraphQL validation

+++ src/modules/offer-request/dto/creation/offer-request-data-initial.input.ts
@@ -1,4 +1,4 @@
-import { Field, InputType, Int } from '@nestjs/graphql';
+import { Field, InputType, Float } from '@nestjs/graphql';
 import { serverTimestamp } from 'firebase/database';
 import { AddressesAnswered, OfferRequestDB } from 'hero24-types';
 import { omitUndefined } from 'src/modules/common/common.utils';
@@ -35,10 +35,10 @@ export class OfferRequestDataInitialInput {
   @Field(() => String, { nullable: true })
   postpaid?: 'no' | 'yes';
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   fixedPrice?: number;
 
-  @Field(() => Int, { nullable: true })
+  @Field(() => Float, { nullable: true })
   fixedDuration?: number;
 
   @Field(() => Boolean, { nullable: true })
